+++ src/windows/CouchLink.ReleaseSecurity/PinnedReleaseKey.cs
@@ -1,0 +1,11 @@
+namespace CouchLink.ReleaseSecurity;
+
+public static class PinnedReleaseKey
+{
+    public const string PublicKeySha256Fingerprint = "28ec51ebeeea7ae970f9c62447868818a51932697f64472ec79e5ab0d906f93c";
+
+    public const string PublicKeyPem = @"-----BEGIN PUBLIC KEY-----
+MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEj/jWCgoLV4Uj8a2cBevVw5MjXlo6
+2KIRyASnqnkpnxdhgdOrWAcS66cn7IcrrG1RcYO4DeUKSDbtV1B1ycPWaA==
+-----END PUBLIC KEY-----";
+}